refactor: extract dev profile helpers
